+++ CHANGES.rst
@@ -37,6 +37,8 @@ astropy.io.ascii
 astropy.io.misc
 ^^^^^^^^^^^^^^^
 
+- Eliminate deprecated compatibility mode when writing ``Table`` metadata to HDF5 format. [#8899]
+
 astropy.io.fits
 ^^^^^^^^^^^^^^^